naming convention cleanup
